backend/tasker/serializers.py
from rest_framework import serializers
from .models import Project, Country, Operation


class ProjectSerializer(serializers.ModelSerializer):
    class Meta:
        model = Project
        fields = (
            'id', 'name', 'note', 'created_at', 'updated_at', 'country', 'country_name'
        )


class CountrySerializer(serializers.ModelSerializer):
    class Meta:
        model = Country
        fields = '__all__'


class OperationSerializer(serializers.ModelSerializer):
    class Meta:
        model = Operation
        fields = '__all__'
