backend/requirements.txt
Django==2.0.7
psycopg2==2.7.5
psycopg2-binary==2.7.5
pytz==2018.5
django-filter==1.1.0
djangorestframework==3.8.2
Markdown==2.6.11